Allow users to opt-in for existing PR description overwrite
Allow users to run the action and overwrite the PR description
even if one already exists.
This will allow for the description to remain consistent with the
PR changes and not require for the users to manually clear out the
PR description in order to get the Action to run.

Requested and fixes #5